apps/backend/api/pyproject.toml
[project]
name = "bovi-api"
version = "0.1.0"
description = "Bovi central API — gateway, persistence (PostgreSQL), and monitoring"
authors = [{ name = "Douwe de Kok", email = "douwedekok@gmail.com" }]
requires-python = ">= 3.12, < 3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.22",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bovi_api"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
